requirements.txt
pymodbus==3.9.2
paho-mqtt==1.6.1
flask==2.3.2
pyyaml==6.0.2
loguru==0.6.0